nfaces/requirements.txt
# System Dependencies (install via apt-get):
# - python3-tk (for legacy GUI support - optional if using Gradio)
# - python3-pil.imagetk (for PIL/Tkinter image integration - optional)
# - python3-pip (for package management)
#
# Install system dependencies:
#   sudo apt-get install -y python3-tk python3-pil.imagetk python3-pip

# Core Python Package Dependencies
numpy>=1.21.0
Pillow>=9.0.0
requests>=2.25.0
opencv-python>=4.5.0

# PostgreSQL and pgvector support
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0

# Modern Web-based GUI (Gradio)
gradio>=4.0.0

# ============================================================================
# Face Embedding Models (Optional - choose based on your needs)
# ============================================================================

# FaceNet - High accuracy deep learning model
# Recommended for most use cases
facenet-pytorch>=2.5.0
torch>=2.0.0
torchvision>=0.15.0

# InsightFace (ArcFace) - State-of-the-art face recognition
# Best accuracy but larger model size
insightface>=0.7.0
onnxruntime>=1.15.0

# DeepFace - Multi-purpose framework (supports multiple models)
# Provides VGGFace2, OpenFace, and other models
deepface>=0.0.79

# ============================================================================
# Installation Notes:
# ============================================================================
#
# FULL INSTALLATION (all models):
#   pip3 install -r requirements.txt
#
# MINIMAL INSTALLATION (statistical model only):
#   pip3 install numpy Pillow requests opencv-python psycopg2-binary python-dotenv gradio
#
# RECOMMENDED INSTALLATION (with FaceNet):
#   pip3 install numpy Pillow requests opencv-python psycopg2-binary python-dotenv gradio facenet-pytorch torch torchvision
#
# If you get "externally-managed-environment" error on some systems:
#   pip3 install --break-system-packages -r requirements.txt
#
# For ARM/Raspberry Pi users:
#   - torch and torchvision may require special ARM builds
#   - Consider using statistical model or lightweight alternatives
#   - InsightFace may have ARM compatibility issues